rename to GitTagData
- hide that it has a read tags AsyncSequence
- set up so that it can be use to write out [TagData] in a future diff (which won't be AsyncSequence)
